feat: refactor card display code
