+++ PocketKit/Sources/PocketKit/Collections/CollectionViewController.swift
@@ -63,12 +63,9 @@ class CollectionViewController: UIViewController {
         return collectionView
     }()
 
+    private let sectionProvider: GridSectionLayoutProvider
     private var metadata: CollectionMetadata?
     private var subscriptions: [AnyCancellable] = []
-
-    private var showThreeColumnLayout: Bool {
-        return self.view.bounds.width >= Constants.minWidthBoundaryForFullColumnLayout && !UIDevice.current.orientation.isPortrait
-    }
 
     private let model: CollectionViewModel
 
@@ -76,6 +73,8 @@ class CollectionViewController: UIViewController {
         model: CollectionViewModel
     ) {
         self.model = model
+
+        sectionProvider = GridSectionLayoutProvider()
 
         super.init(nibName: nil, bundle: nil)
 
@@ -207,11 +206,6 @@ private extension CollectionViewController {
         /// Height that centers the error section so that it appears approximately in the middle
         static let errorSectionHeight: CGFloat = 0.65
         static let errorSectionInsets = NSDirectionalEdgeInsets(top: 0, leading: 20, bottom: 0, trailing: 20)
-
-        static let itemPadding = NSDirectionalEdgeInsets(top: 8, leading: 0, bottom: 8, trailing: 0)
-        static let interItemSpacing: CGFloat = 16
-        /// Minimum width length to qualify for a full (3 column) grid layout on iPad
-        static let minWidthBoundaryForFullColumnLayout: CGFloat = 800
     }
 
     func section(for index: Int, environment: NSCollectionLayoutEnvironment) -> NSCollectionLayoutSection? {
@@ -261,15 +255,10 @@ private extension CollectionViewController {
 
             return section
         case .collection(let collection):
-            let width = environment.container.effectiveContentSize.width
-            let margin: CGFloat = environment.traitCollection.shouldUseWideLayout() ? Margins.iPadNormal.rawValue : Margins.normal.rawValue
             let stories = collection.stories?.compactMap { $0 as? CollectionStory } ?? []
+            let viewModels = stories.compactMap { model.storyViewModel(for: $0) }
 
-            if environment.traitCollection.shouldUseWideLayout() {
-                return collectionSectionForWideLayout(with: stories, width: width, margin: margin)
-            } else {
-                return collectionSectionForCompactLayout(with: stories, width: width, margin: margin)
-            }
+            return sectionProvider.gridSection(for: viewModels, with: environment, and: view)
         case .error:
             let section = NSCollectionLayoutSection(
                 group: .vertical(
@@ -292,134 +281,6 @@ private extension CollectionViewController {
         default:
             return .empty()
         }
-    }
-
-    /// Determines the stories layout in the native collection view on iPad mode with regular horizontal size class (including split view). Number of columns for the grid layout depends on device orientation and the view's width length. If view's width is less than 800, show 2 col, otherwise show 2 col if it is in portrait and 3 col if it is in landscape mode.
-    /// - Parameters:
-    ///   - stories: list of stories associated with collection to display
-    ///   - width: width that the section occupies
-    ///   - margin: padding adding to the side of the section
-    /// - Returns: section layout for stories on iPad view and regular horizontal size class
-    private func collectionSectionForWideLayout(with stories: [CollectionStory], width: CGFloat, margin: CGFloat) -> NSCollectionLayoutSection {
-        let numberOfColumns: CGFloat = showThreeColumnLayout ? 3 : 2
-        let recommendationsHeight: [CGFloat] = getRecommendationHeights(with: stories, width: width, margin: margin, numberOfColumns: numberOfColumns)
-
-        /// Retrieves max height for each row and returns an array of row heights
-        let totalGroupHeights = recommendationsHeight.getMaxHeightForRow(of: Int(numberOfColumns))
-
-        let components = createComponentsForWideLayout(with: stories, numberOfColumns: numberOfColumns, totalGroupHeights: totalGroupHeights)
-        return createSectionFromGroup(with: components, and: margin)
-    }
-
-    /// Create components needed for wide section layout
-    /// - Parameters:
-    ///   - recommendations: list of stories to display to the user
-    ///   - width: width that the section occupies
-    ///   - sideMargin: margin for the section layout
-    /// - Returns: tuple of total height and list of items used to determine section layout
-    private func createComponentsForWideLayout(with stories: [CollectionStory], numberOfColumns: CGFloat, totalGroupHeights: [CGFloat]) -> (CGFloat, [NSCollectionLayoutGroup]) {
-        let numberOfRows = (CGFloat(stories.count) / numberOfColumns).rounded(.up)
-        let item = NSCollectionLayoutItem(
-            layoutSize: NSCollectionLayoutSize(
-                widthDimension: .fractionalWidth(1 / numberOfColumns),
-                heightDimension: .fractionalHeight(1)
-            )
-        )
-        item.contentInsets = Constants.itemPadding
-
-        return (0..<Int(numberOfRows)).reduce((CGFloat(0), [NSCollectionLayoutGroup]())) { result, rowIndex in
-            let currentHeight = result.0
-            guard let height = totalGroupHeights[safe: rowIndex] else { return result }
-            var groups = result.1
-            let group = NSCollectionLayoutGroup.horizontal(
-                layoutSize: .init(
-                    widthDimension: .fractionalWidth(1),
-                    heightDimension: .absolute(height)
-                ),
-                repeatingSubitem: item,
-                count: Int(numberOfColumns)
-            )
-            group.interItemSpacing = .fixed(Constants.interItemSpacing)
-            groups.append(group)
-            return (currentHeight + height, groups)
-        }
-    }
-
-    /// Get a list of heights for each story
-    /// - Parameters:
-    ///   - recommendations: list of stories
-    ///   - width: width that the section occupies
-    ///   - sideMargin: padding adding to the side of the section
-    ///   - numberOfColumns: number of columns layout should have
-    /// - Returns: return a list of heights for all the stories
-    private func getRecommendationHeights(with stories: [CollectionStory], width: CGFloat, margin: CGFloat, numberOfColumns: CGFloat) -> [CGFloat] {
-        return stories.map { story in
-            return RecommendationCell.fullHeight(viewModel: model.storyViewModel(for: story), availableWidth: width / numberOfColumns - (margin * 2)) + margin
-        }
-    }
-
-    /// Determines the stories layout for the native collection view on iPhone mode (single column layout)
-    /// - Parameters:
-    ///   - stories: list of stories associated with collection to display
-    ///   - width: width that the section occupies
-    ///   - margin: padding adding to the side of the section
-    /// - Returns: section layout for compact (i.e. iPhone mode)
-    private func collectionSectionForCompactLayout(with stories: [CollectionStory], width: CGFloat, margin: CGFloat) -> NSCollectionLayoutSection {
-        let components = createComponentsForCompact(with: stories, width: width, margin: margin)
-        return createSectionFromGroup(with: components, and: margin)
-    }
-
-    /// Create components needed for compact section layout
-    /// - Parameters:
-    ///   - recommendations: list of stories to display to the user
-    ///   - width: width that the section occupies
-    ///   - sideMargin: margin for the section layout
-    /// - Returns: tuple of total height and list of items used to determine section layout
-    private func createComponentsForCompact(with stories: [CollectionStory], width: CGFloat, margin: CGFloat) -> (CGFloat, [NSCollectionLayoutItem]) {
-        return stories.reduce((CGFloat(0), [NSCollectionLayoutItem]())) { result, story in
-            let currentHeight = result.0
-
-            let height = RecommendationCell.fullHeight(
-                viewModel: model.storyViewModel(for: story),
-                availableWidth: width - (margin * 2)
-            ) + margin
-
-            var items = result.1
-            let item = NSCollectionLayoutItem(
-                layoutSize: NSCollectionLayoutSize(
-                    widthDimension: .fractionalWidth(1),
-                    heightDimension: .absolute(height)
-                )
-            )
-            item.contentInsets = Constants.itemPadding
-
-            items.append(item)
-
-            return (currentHeight + height, items)
-        }
-    }
-
-    /// Creates section from group layout
-    /// - Parameters:
-    ///   - components: components that consist of height and list of items for the group
-    ///   - margin: margin for the section layout
-    /// - Returns: section layout
-    private func createSectionFromGroup(with components: (CGFloat, [NSCollectionLayoutItem]), and margin: CGFloat) -> NSCollectionLayoutSection {
-        let group = NSCollectionLayoutGroup.vertical(
-            layoutSize: .init(
-                widthDimension: .fractionalWidth(1),
-                heightDimension: .absolute(components.0)
-            ),
-            subitems: components.1
-        )
-        let section = NSCollectionLayoutSection(group: group)
-        section.contentInsets = NSDirectionalEdgeInsets(
-            top: 0,
-            leading: margin,
-            bottom: 0,
-            trailing: margin
-        )
-        return section
     }
 
     func cell(
