+++ PocketKit/Sources/SharedPocketKit/Array+Extension.swift
@@ -13,24 +13,3 @@ extension Array {
         return self[index]
     }
 }
-
-extension Array where Element == CGFloat {
-    /// Retrieves a list of heights for each row
-    /// - Parameter size: number of items in a row
-    /// - Returns: list of heights for each row
-    public func getMaxHeightForRow(of size: Int) -> [CGFloat] {
-        guard size > 0 else { return [] }
-        var maxHeights: [CGFloat] = []
-
-        for index in stride(from: 0, to: count, by: size) {
-            let lastChunkIndex = Swift.min(index + size, count)
-            let row = self[index..<lastChunkIndex]
-
-            if let maxValue = row.sorted().last {
-                maxHeights.append(maxValue)
-            }
-        }
-
-        return maxHeights
-    }
-}
